fix: renamed folders and code cleaning
