+++ product-backend/src/main/java/org/kainos/ea/controller/AuthController.java
@@ -7,16 +7,18 @@ import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
-import org.kainos.ea.api.AuthService;
-import org.kainos.ea.cli.User;
+import org.kainos.ea.db.AuthDao;
 import org.kainos.ea.exception.FailedToRegisterException;
 import org.kainos.ea.exception.InvalidEmailException;
 import org.kainos.ea.exception.InvalidPasswordException;
+import org.kainos.ea.model.User;
+import org.kainos.ea.service.AuthService;
+import org.kainos.ea.service.AuthValidator;
 
 @Tag(name = "Authorisation")
 @Path("/api")
 public class AuthController {
-    private AuthService authService = new AuthService();
+    private final AuthService authService = new AuthService(new AuthDao(), new AuthValidator());
 
     @POST
     @Path("/auth/register")
