+++ product-backend/src/main/java/org/kainos/ea/model/User.java
@@ -1,4 +1,4 @@
-package org.kainos.ea.cli;
+package org.kainos.ea.model;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -7,6 +7,16 @@ public class User {
     private String email;
     private String password;
     private int roleId;
+
+    @JsonCreator
+    public User(
+            @JsonProperty("email") String email,
+            @JsonProperty("password") String password,
+            @JsonProperty("roleId") int roleId) {
+        this.email = email;
+        this.password = password;
+        this.roleId = roleId;
+    }
 
     public String getEmail() {
         return email;
@@ -29,16 +39,6 @@ public class User {
     }
 
     public void setRoleId(int roleId) {
-        this.roleId = roleId;
-    }
-
-    @JsonCreator
-    public User(
-            @JsonProperty("email") String email,
-            @JsonProperty("password") String password,
-            @JsonProperty("roleId") int roleId){
-        this.email = email;
-        this.password = password;
         this.roleId = roleId;
     }
 }
